Notification System
The notifications are pulled from the back end now, and display and can be cleared. When the database is implemented, these will be way cooler than they currently are. There's also just some light fixes in other files that didn't make it to my branch. The real important files are the notification ones. Next push should be unit testing.